refactor(domain): move FileContent to value objects
